requirements.txt
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
Django==2.2.4
django-filter==2.2.0
djangorestframework==3.9.3
djangorestframework-stubs==0.4.2
gunicorn==19.9.0
idna==2.8
Markdown==3.1.1
psycopg2-binary==2.8.2
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
